% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py_simple.R
\name{as.pinyin}
\alias{as.pinyin}
\title{Convert strings of Chinese characters into Pinyin.}
\usage{
as.pinyin(char, sep = "_", other_replace = NULL, dic = pinyin::pinyin2)
}
\arguments{
\item{char}{\emph{Requrired} A input character string with chinese characters.}

\item{sep}{character  The seperator between the converted pinyin. Default: '_'}

\item{other_replace}{character  The replacement for non-Chinese characters}

\item{dic}{The preloaded dictionary for conversion. Default: "pinyin::pinyin2"}
}
\value{
A character vector of the converted pinyin string.
}
\description{
Convert strings of Chinese characters into Pinyin.
}
\examples{
py("hello word")

}
