% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py.R
\name{bookdown2py}
\alias{bookdown2py}
\title{Convert the Chinese headers of bookdown .Rmd files into Pinyin}
\usage{
bookdown2py(folder = "py", remove_curly_bracket = TRUE, other_replace = NULL, dic = NA)
}
\arguments{
\item{folder}{character. The folder in which the files are to be converted.}

\item{remove_curly_bracket}{logical. Whether to remove existing curly brackets in the headers.}

\item{other_replace}{NULL or character. Define how to convert non-Chinese characters in mychar. NULL means 'let it be'.}

\item{dic}{See \code{help(pinzin)}.}
}
\value{
new .Rmd files with Pinyin headers.
}
\description{
Convert the Chinese headers of bookdown .Rmd files into Pinyin
}
\examples{
mydir <- paste0(tempdir(), "/py")
dir.create(mydir)
file.create(paste0(mydir, "/test.txt"))
writeLines(text = "# test\\n", paste0(mydir, "/test.txt"))
bookdown2py(mydir)
}
