% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py.R
\name{file.rename2py}
\alias{file.rename2py}
\title{Rename files according to a given dictionary}
\usage{
file.rename2py(folder = "py", dic = NA)
}
\arguments{
\item{folder}{character. The folder in which the files are to be renamed.}

\item{dic}{See \code{help(pinyin)}.}
}
\value{
files with new names.
}
\description{
Rename files according to a given dictionary
}
\examples{
mydir <- paste0(tempdir(), "/py")
dir.create(mydir)
file.create(paste0(mydir, "/test.txt"))
file.rename2(mydir)
}
