% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py.R
\name{file2py}
\alias{file2py}
\title{Convert the characters in an entire files according to a given dictionary}
\usage{
file2py(folder = "py", backup = TRUE, sep = " ", other_replace = NULL, 
    encoding = "UTF-8", dic = NA)
}
\arguments{
\item{folder}{character. The folder in which the files are to be converted.}

\item{backup}{logical. Whether the original files should be saved as backups.}

\item{sep}{character. Seperation between the converted pinyin.}

\item{other_replace}{NULL or character. Define how to convert non-Chinese characters in mychar. NULL means 'let it be'.}

\item{encoding}{character. The encoding of the input files. 'UTF-8' by default.}

\item{dic}{See \code{help(pinzin)}.}
}
\value{
files converted to Pinyin.
}
\description{
Convert the characters in an entire files according to a given dictionary
}
\examples{
file2py(dic = NA)
}
