% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/four_corner.R
\name{four_corner}
\alias{four_corner}
\title{Get the four-corner code of a string}
\usage{
four_corner(mychar = "", sep = "_", nonezh_replace = NULL, fc = fclib())
}
\arguments{
\item{mychar}{character. A Chinese character or string to convert to four-corner codes.}

\item{sep}{character. Seperation between the converted four-corner codes.}

\item{nonezh_replace}{NULL or character. Define how to convert non-Chinese characters in mychar. NULL means 'let it be'.}

\item{fc}{A four-corner library}
}
\value{
the four-corner codes
}
\description{
Get the four-corner code of a string
}
\examples{
four_corner()
}
