% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{insert}
\alias{insert}
\title{Title insert a character(s) before all lattin letters in a string}
\usage{
insert(oldchar, insertchar = ",")
}
\arguments{
\item{oldchar}{the old string}

\item{insertchar}{the character(s) to insert}
}
\value{
a new string
}
\description{
Title insert a character(s) before all lattin letters in a string
}
\examples{
insert(" abc")
}
