% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py.R
\name{load_dic}
\alias{load_dic}
\title{Load a customized dictionary.}
\usage{
load_dic(dic_file = NA, select = 1)
}
\arguments{
\item{dic_file}{The path of a dictionary file.}

\item{select}{The option to choose from the dictionary.}
}
\value{
A dictionary
}
\description{
Load a customized dictionary.
}
\examples{
load_dic()
}
