% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinyin.R
\name{pinyin}
\alias{pinyin}
\title{Convert Chinese characters into Pinyin.}
\usage{
pinyin(mychar = "", sep = "_", nonezh_replace = NULL, py = NA, method = c("quanpin", 
    "tone", "toneless"), multi = FALSE, only_first_letter = FALSE, dic = c("zh", 
    "zh2"))
}
\arguments{
\item{mychar}{character. A Chinese character or string to convert to pinyin}

\item{sep}{character. Seperation between the converted pinyin.}

\item{nonezh_replace}{NULL or character. Define how to convert non-Chinese characters in mychar. NULL means 'let it be'.}

\item{py}{the preloaded pinyin library using the \code{pylib()} function.}

\item{method}{character. Only valid when py == NA. The value can be:
\itemize{
\item 'quanpin', for the standard form of pinyin (tones above letters),
\item 'tone', for tones expressed with numbers,
\item 'toneless', without tones
}}

\item{multi}{logical. Only valid when py == NA. Whether display multiple pronounciations of a Chinese character or only the first pronounciation.}

\item{only_first_letter}{logical. Only valid when py == NA. Wheter only the first letter in pinyin.}

\item{dic}{character. Only valid when py == NA. Choose the dictionary.}
}
\value{
pinyin of the given Chinese character.
}
\description{
Convert Chinese characters into Pinyin.
}
\examples{
pinyin("test")
}
