% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pinyin2}
\alias{pinyin2}
\title{The Dictionary to Convert Chinese Characters to Pinyin.}
\format{A data frame with 20520 rows and 2 variables:
\describe{
\item{key}{a chinese character}
\item{value}{a pinyin}
...
}}
\usage{
pinyin2
}
\description{
A dataset containing the Chinese characters and the corresponding pinyin, Source: A friend of dapeng
}
\keyword{datasets}
