% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py.R
\name{py_single}
\alias{py_single}
\title{Convert a string of Chinese characters into Pinyin.}
\usage{
py_single(char = "", sep = "_", other_replace = NULL, dic = pydic())
}
\arguments{
\item{char}{character. A Chinese character or string to convert to pinyin}

\item{sep}{character. Seperation between the converted pinyin.}

\item{other_replace}{NULL or character. Define how to convert non-Chinese characters in mychar. NULL means 'let it be'.}

\item{dic}{the preloaded pinyin library using the \code{pylib()} function.}
}
\value{
pinyin of the given Chinese string.
}
\description{
Convert a string of Chinese characters into Pinyin.
}
