% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinyin.R
\name{pylib}
\alias{pylib}
\title{A Pinyin library}
\usage{
pylib(method = c("quanpin", "tone", "toneless"), multi = FALSE, 
    only_first_letter = FALSE, dic = c("zh", "zh2"))
}
\arguments{
\item{method}{character. The value can be:
\itemize{
\item 'quanpin', for the standard form of pinyin (tones above letters),
\item 'tone', for tones expressed with numbers,
\item 'toneless', without tones
}}

\item{multi}{logical. Whether display multiple pronounciations of a Chinese character or only the first pronounciation.}

\item{only_first_letter}{logical. Wheter only the first letter in pinyin.}

\item{dic}{character. Choose the dictionary.}
}
\value{
character. a Pinyin library.
}
\description{
A Pinyin library
}
\examples{
pylib()
}
