% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{strsplit2}
\alias{strsplit2}
\title{split a string and extract the \emph{n}th string}
\usage{
strsplit2(x, sep = " ", nth = 1)
}
\arguments{
\item{x}{The string to split}

\item{sep}{The separated character}

\item{nth}{The \emph{n}th string to extract}
}
\value{
a new string
}
\description{
split a string and extract the \emph{n}th string
}
\examples{
strsplit2("char1 char2")
}
