% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wubi.R
\name{wubi}
\alias{wubi}
\title{Convert Chinese strings to wubi code (based on radicals).}
\usage{
wubi(Chin.strs = NA, sep = "_", parallel = FALSE)
}
\arguments{
\item{Chin.strs}{The string need to be converted}

\item{sep}{Character used to seperate different characters}

\item{parallel}{Whether or not use parallel calculation}
}
\value{
wubi code of \code{Chin.str}.
}
\description{
Convert Chinese strings to wubi code (based on radicals).
}
\examples{
data(WBlib)
wubi()
}
