% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{zh2py}
\alias{zh2py}
\title{Concert Chinese characters into Pinyin.}
\usage{
zh2py(mychar, method = c("quanpin", "tone", "toneless")[1], sep = "_", 
    nonezh_replace = NULL, only_first_letter = FALSE)
}
\arguments{
\item{mychar}{character. A Chinese character or string to convert to pinyin}

\item{method}{character. The value can be:
- 'quanpin', when '汉字' is to be converted to 'hàn zì',
- 'tone', when '汉字' is to be converted to han4 zi4,
- 'toneless', when '汉字' is to be converted to 'han zi'}

\item{sep}{character. Seperation between the converted pinyin.}

\item{nonezh_replace}{NULL or character. Define how to convert non-Chinese characters in mychar. NULL means 'let it be'.}

\item{only_first_letter}{logical. Wheter only the first letter in pinyin.}
}
\value{
pinyin of the given Chinese character.
}
\description{
Concert Chinese characters into Pinyin.
}
\examples{
zh2py()
}
