% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{bookdown2py}
\alias{bookdown2py}
\title{Convert the Chinese headers of bookdown .Rmd files into Pinyin}
\usage{
bookdown2py(folder = "py", remove_curly_bracket = TRUE)
}
\arguments{
\item{folder}{character. The folder in which the files are to be converted.}

\item{remove_curly_bracket}{logical. Whether to remove existing curly brackets in the headers.}
}
\value{
new .Rmd files with Pinyin headers.
}
\description{
Convert the Chinese headers of bookdown .Rmd files into Pinyin
}
\examples{
bookdown2py()
}
