% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinyin.R
\name{file.rename2py}
\alias{file.rename2py}
\title{Rename files with Chinese characters to pinyin}
\usage{
file.rename2py(folder = "py", py = NA, dic = c("zh", "zh2"))
}
\arguments{
\item{folder}{character. The folder in which the files are to be renamed.}

\item{py}{See \code{help(pinyin)}.}

\item{dic}{See \code{help(pinzin)}.}
}
\value{
files with new names.
}
\description{
Rename files with Chinese characters to pinyin
}
\examples{
mydir <- paste0(tempdir(), "/py")
dir.create(mydir)
file.create(paste0(mydir, "/test.txt"))
file.rename2py(mydir)
}
