% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{file.rename2py}
\alias{file.rename2py}
\title{Renames files with Chinese characters to pinyin}
\usage{
file.rename2py(mydir = "/")
}
\arguments{
\item{mydir}{character. The folder in which the files are to be renamed.}
}
\value{
files with new names.
}
\description{
Renames files with Chinese characters to pinyin
}
