% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{pylib}
\alias{pylib}
\title{A Pinyin library}
\usage{
pylib(method = c("quanpin", "tone", "toneless")[1], first = TRUE, 
    only_first_letter = FALSE)
}
\arguments{
\item{method}{character. The value can be:
- 'quanpin', when '汉字' is to be converted to 'hàn zì',
- 'tone', when '汉字' is to be converted to han4 zi4,
- 'toneless', when '汉字' is to be converted to 'han zi'}

\item{first}{logical. Whether only display the first pronounciation of multiple pronounciations of a Chinese character.}

\item{only_first_letter}{logical. Wheter only the first letter in pinyin.}
}
\value{
a Pinyin library.
}
\description{
A Pinyin library
}
\examples{
pylib()
pylib(first = FALSE)
}
